myInfo data for both peripheral and node info model fixed data loading and cleaned up models

+++ MeshtasticClient/Views/Nodes/NodeList.swift
@@ -6,7 +6,7 @@
 //
 
 // Abstract:
-//  A view showing a list of devices that have been seen on the mesh network
+//  A view showing a list of devices that have been seen on the mesh network from the perspective of the connected device.
 
 import SwiftUI
 
@@ -26,29 +26,56 @@ struct NodeList: View {
         NavigationView {
            
             List {
-                Toggle(isOn: $showLocationOnly) {
-                    Text("Nodes with location only")
+
+                if meshData.nodes.count == 0 {
+                    Text("Scan for Radios").font(.largeTitle).listRowSeparator(.hidden)
+                    Text("No LoRa Mesh Nodes Found").font(.title2).listRowSeparator(.hidden)
+                    Text("Go to the bluetooth section in the bottom right menu and click the Start Scanning button to scan for nearby radios and find your Meshtastic device. Make sure your device is powered on and near your phone or tablet.")
+                        .font(.body)
+                        .listRowSeparator(.hidden)
+                    Text("Once the device shows under Available Devices touch the device you want to connect to and it will pull node information over BLE and populate the node list and mesh map in the Meshtastic app.")
+                        .listRowSeparator(.hidden)
+                    Text("Views with bluetooth functionality will show an indicator in the upper right hand corner show if bluetooth is on, and if a device is connected.")
+                        .listRowSeparator(.hidden)
+                    Spacer().listRowSeparator(.hidden)
+                    //Button(action: {}) {
+                    //    Text("Get Started")
+                    //        .font(.title)
+                    //        .frame(maxWidth: 300)
+                    //}
+                    //.buttonStyle(.borderedProminent)
+                    //.buttonBorderShape(.automatic)
+                    //.controlSize(.large)
                 }
-                ForEach(filteredDevices.sorted(by: { $0.lastHeard > $1.lastHeard })) { node in
-                    NavigationLink(destination: NodeDetail(node: node)) {
-                        NodeRow(node: node, index : 0)
-    
+                else {
+                    Toggle(isOn: $showLocationOnly) {
+                        Text("Nodes with location only")
                     }
-                    .swipeActions {
-                        Button {
-                            let nodeIndex = meshData.nodes.firstIndex(where: { $0.id == node.id })
-                            meshData.nodes.remove(at: nodeIndex!)
-                            meshData.save()
-                        } label: {
-                            VStack {
-                                Label("Delete from app", systemImage: "trash")
-                            }
+                    ForEach(filteredDevices.sorted(by: { $0.lastHeard > $1.lastHeard })) { node in
+                        NavigationLink(destination: NodeDetail(node: node)) {
+                            NodeRow(node: node, index : 0)
+        
                         }
-                        .tint(.red)
+                        .swipeActions {
+                            Button {
+                                let nodeIndex = meshData.nodes.firstIndex(where: { $0.id == node.id })
+                                meshData.nodes.remove(at: nodeIndex!)
+                                meshData.save()
+                            } label: {
+                                VStack {
+                                    Label("Delete from app", systemImage: "trash")
+                                }
+                            }
+                            .tint(.red)
+                        }
                     }
                 }
              }
             .navigationTitle("All Nodes")
+        }
+        .ignoresSafeArea(.all, edges: [.leading, .trailing])
+        .onAppear{
+            meshData.load()
         }
     }
 }
